adds full support for fuzzy string matching
